Bug: add confirmation page to prevent Microsoft Safe Links from consuming magic link token. #315
